refactor(canvas): extract chat mention sources
