+++ apps/canvas-workspace/src/renderer/src/agent-chat/mentions/roleMentionItems.test.ts
