+++ apps/canvas-workspace/src/renderer/src/types/agent-chat.ts
@@ -2,6 +2,8 @@ import type {
   AgentChatMessage,
   AgentChatMcpApp,
   AgentClarificationRequest,
+  AgentContextDomSelectionRef,
+  AgentContextTabRef,
   AgentRequestContext,
   AgentScope,
   AgentScopeRef,
@@ -10,6 +12,7 @@ import type {
   CrossWorkspaceSessionGroup,
   SessionSearchHit,
 } from '../../../shared/agent-chat';
+import type { CanvasNode } from './canvas';
 import type {
   RoleTurnEndEvent,
   RoleTurnRoleRef,
@@ -28,6 +31,41 @@ export interface AgentNewSessionResult {
   activeSessionId?: string | null;
   code?: string;
   error?: string;
+}
+
+export interface WorkspaceOption {
+  id: string;
+  name: string;
+}
+
+export interface MentionItem {
+  type: 'node' | 'file' | 'folder' | 'workspace' | 'skill' | 'plugin' | 'tag' | 'session' | 'dom' | 'tab' | 'role';
+  label: string;
+  /** For type === 'plugin': stable plugin-market listing id. */
+  pluginId?: string;
+  /** Client-owned brand key; Agent Plugins itself has no portable icon field. */
+  pluginIconKey?: string;
+  /** For type === 'role': the chat persona's id from the role library. */
+  roleId?: string;
+  /** For type === 'role': the persona's accent color (popup icon tint). */
+  roleColor?: string;
+  nodeType?: CanvasNode['type'];
+  /** For type === 'node': the canvas node id, used to focus it when clicked. */
+  nodeId?: string;
+  path?: string;
+  workspaceId?: string;
+  /** For type === 'tag': workspaces the tag occurs in (global assistant). */
+  workspaceIds?: string[];
+  /** Extra context shown in the popup row (e.g. skill detail or node workspace). */
+  description?: string;
+  /** For type === 'session': the referenced chat session id. */
+  sessionId?: string;
+  /** For type === 'session': index of the first message matching the query. */
+  messageIndex?: number;
+  /** For type === 'dom': selected iframe/webview DOM element context. */
+  domSelection?: AgentContextDomSelectionRef;
+  /** For type === 'tab': the referenced right-dock tab. */
+  tab?: AgentContextTabRef;
 }
 
 /** Progress shown while a multi-role conversation advances through speakers. */
